+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * fix position of textarea in multi line command (probably will affect IME or Emoji picker)
 * make jump to bottom on click only when terminal is not enabled [#596](https://github.com/jcubic/jquery.terminal/issues/596)
 * fix init cmd plugin without terminal
+* fix underscore cut off in Firefox (visible in signature)
 
 ## 2.17.5
 ### Bugfix